refactor: :recycle: improve shelters list table folders structure
