+++ src/app/user/shelters/[id]/edit/page.tsx
@@ -4,14 +4,21 @@ import { notFound } from "next/navigation";
 import { api } from "~/trpc/react";
 import { FormEditRegister } from "~/components/shelter/form-edit-register";
 import { Skeleton } from "~/components/ui/skeleton";
+import { toast } from "sonner";
 
 export default function ShelterPage({ params }: { params: { id: number } }) {
   const { data, isLoading, error } = api.shelter.findUserShelterById.useQuery({
     id: Number(params.id),
   });
 
-  if (error?.data?.httpStatus === 404 || (!data && !isLoading)) {
+  if (error?.data?.httpStatus === 404 && !isLoading) {
     notFound();
+  }
+
+  if (!data && !isLoading) {
+    toast.error(
+      "Erro ao carregar os dados abrigo. Tente novamente mais tarde.",
+    );
   }
 
   if (isLoading) {
